refactor: create cose utilities module, move header_util.rs there
