+++ src/token/cose/encrypted/encrypt0/tests.rs
@@ -15,7 +15,7 @@ use rstest::rstest;
 
 use crate::token::cose::encrypted::encrypt0::{CoseEncrypt0BuilderExt, CoseEncrypt0Ext};
 use crate::token::cose::encrypted::EncryptCryptoBackend;
-use crate::token::cose::header_util::HeaderBuilderExt;
+use crate::token::cose::header::HeaderBuilderExt;
 use crate::token::cose::test_helper::{
     apply_attribute_failures, apply_header_failures, perform_cose_reference_output_test,
     perform_cose_self_signed_test, serialize_cose_with_failures, CoseStructTestHelper, TestCase,
