+++ src/token/cose/maced/mac0/tests.rs
@@ -15,7 +15,7 @@ use coset::iana::Algorithm;
 use coset::{CoseError, CoseKey, CoseKeyBuilder, CoseMac0, CoseMac0Builder, Header};
 use rstest::rstest;
 
-use crate::token::cose::header_util::determine_algorithm;
+use crate::token::cose::header::determine_algorithm;
 use crate::token::cose::maced::mac0::{CoseMac0BuilderExt, CoseMac0Ext};
 use crate::token::cose::maced::MacCryptoBackend;
 use crate::token::cose::test_helper::{
