+++ src/token/cose/signed/mod.rs
@@ -18,7 +18,7 @@ pub use sign1::{CoseSign1BuilderExt, CoseSign1Ext};
 
 use crate::error::CoseCipherError;
 use crate::token::cose::key::{CoseEc2Key, CoseParsedKey, KeyProvider};
-use crate::token::cose::{header_util, key, CryptoBackend};
+use crate::token::cose::{header, key, CryptoBackend};
 
 mod sign;
 mod sign1;
@@ -183,7 +183,7 @@ fn try_sign<B: SignCryptoBackend, CKP: KeyProvider>(
     unprotected: Option<&Header>,
     payload: &[u8],
 ) -> Result<Vec<u8>, CoseCipherError<B::Error>> {
-    header_util::try_cose_crypto_operation(
+    header::try_cose_crypto_operation(
         key_provider,
         protected,
         unprotected,
@@ -227,7 +227,7 @@ fn try_verify<B: SignCryptoBackend, CKP: KeyProvider>(
     signature: &[u8],
     toverify: &[u8],
 ) -> Result<(), CoseCipherError<B::Error>> {
-    header_util::try_cose_crypto_operation(
+    header::try_cose_crypto_operation(
         key_provider,
         Some(protected),
         Some(unprotected),
