+++ src/token/cose/util/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod header;
+pub(crate) mod ops;
+pub(crate) mod symm;
